+++ Documentation/markdown/ReleaseNotes.md
@@ -13,6 +13,8 @@
  
   * Supporting IfcSeventhOrderPolynomialSpiral ([#552](https://github.com/tumcms/Open-Infra-Platform/pull/552))
 
+  * Supporting IfcThirdOrderPolynomialSpiral ([#553](https://github.com/tumcms/Open-Infra-Platform/pull/553))
+
 ## 4.0.0 Release Notes
 
 *April 26th 2022*